feat: implement new Page2Info component with enhanced navigation and styling
- Introduced Page2Info component to manage a paginated view with scrollable pages.
- Replaced previous ViewPager implementation with a more streamlined approach using refs for page management.
